first commit - move manager page
