+++ apps/web/src/components/Router.tsx
@@ -40,7 +40,7 @@ const createRoute = (locale: Locales) =>
                   },
                   lazy: () =>
                     import(
-                      "../pages/ManagerDashboardPage/ManagerDashboardPage"
+                      "../pages/Manager/ManagerDashboardPage/ManagerDashboardPage"
                     ),
                 },
               ],
