+++ apps/web/src/pages/Manager/ManagerDashboardPage/messages.ts
@@ -1,0 +1,17 @@
+import { defineMessages } from "react-intl";
+
+const messages = defineMessages({
+  pageTitle: {
+    defaultMessage: "Manager dashboard",
+    id: "Q914Cl",
+    description: "Title for the 'Manager dashboard' page",
+  },
+  pageSubtitle: {
+    defaultMessage:
+      "Manage your talent requests and curate your manager profile.",
+    id: "Gw0uTQ",
+    description: "Subtitle for the 'Manager dashboard' page",
+  },
+});
+
+export default messages;
