+++ apps/web/src/pages/Manager/ManagerRequestHistoryPage/messages.ts
@@ -1,0 +1,22 @@
+import { defineMessages } from "react-intl";
+
+const messages = defineMessages({
+  longPageTitle: {
+    defaultMessage: "Your talent requests",
+    id: "Z96+gh",
+    description: "Long title for the 'manager request history' page",
+  },
+  shortPageTitle: {
+    defaultMessage: "Talent requests",
+    id: "hZg6UM",
+    description: "Short title for the 'manager request history' page",
+  },
+  pageSubtitle: {
+    defaultMessage:
+      "View a history of your talent requests, including their status and details.",
+    id: "diuFiP",
+    description: "Subtitle for the 'manager request history' page",
+  },
+});
+
+export default messages;
